# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "short-links service API via FLASK",
    "version": "1.0.0",
    "title": "short-links",
    "contact": {
      "email": "cool.selutin99@yandex.ru"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "Short Link Requests",
      "description": "API for short link service"
    }
  ],
  "paths": {
    "/long_to_short": {
      "post": {
        "tags": [
          "Short Link Requests"
        ],
        "summary": "Create a new short link",
        "requestBody": {
          "description": "Short Link Post Objects",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/create_short_link_request"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/create_short_link_response"
            }
          },
          "400": {
            "description": "Failed. Bad request data."
          }
        }
      }
    },
    "/{short_postfix}": {
      "parameters": [
        {
          "name": "short_postfix",
          "in": "path",
          "required": true,
          "description": "Short postfix of the link that we want to match",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Short Link Requests"
        ],
        "summary": "Redirect to original url with given short postfix",
        "parameters": [
          {
            "in": "path",
            "name": "short_postfix",
            "required": true,
            "description": "Short postfix",
            "schema": {
              "$ref": "#/components/schemas/short_postfix"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to original link successfully"
          },
          "400": {
            "description": "Failed. Incorrect short postfix."
          },
          "404": {
            "description": "Failed. Link by short postfix not found."
          }
        }
      }
    },
    "/statistics/{short_postfix}": {
      "parameters": [
        {
          "name": "short_postfix",
          "in": "path",
          "required": true,
          "description": "Short postfix of the link that we want to match",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Short Link Requests"
        ],
        "summary": "Get statistics of visits by short postfix",
        "parameters": [
          {
            "in": "path",
            "name": "short_postfix",
            "required": true,
            "description": "Short postfix",
            "schema": {
              "$ref": "#/components/schemas/short_postfix"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/statistics"
            }
          },
          "400": {
            "description": "Failed. Incorrect short postfix."
          },
          "404": {
            "description": "Failed. Link by short postfix not found."
          }
        }
      }
    }
  },

  "components": {
    "schemas": {
      "short_postfix": {
        "properties": {
          "short_postfix": {
            "type": "string",
            "example": "ghpzy"
          }
        }
      },
      "statistics": {
        "type": "object",
        "properties": {
          "count": {
            "type": "int",
            "format": "int",
            "example": "14"
          }
        }
      },
      "create_short_link_request": {
        "type": "object",
        "properties": {
          "long_url": {
            "type": "string",
            "format": "url",
            "example": "https://github.com/Sean-Bradley/Seans-Python3-Flask-Rest-Boilerplate/blob/master/static/swagger.json"
          }
        }
      },
      "create_short_link_response": {
        "type": "object",
        "properties": {
          "short_link": {
            "type": "string",
            "format": "url",
            "example": "https://127.0.0.1:5000/ghpzy"
          }
        }
      }
    }
  }
}